+++ ain/tolva/nut/Nut.java
@@ -21,7 +21,9 @@ import java.util.Stack;
 import javax.swing.ImageIcon;
 
 import ain.tolva.nut.backend.*;
-import ain.tolva.nut.exceptions.*;
+import ain.tolva.nut.backend.exceptions.ErrorLog;
+import ain.tolva.nut.backend.exceptions.NoTrayAccessException;
+import ain.tolva.nut.backend.plugins.AddPlugin;
 import ain.tolva.nut.plugin.*;
