+++ ain/tolva/nut/backend/alerts/Alerter.java
@@ -1,0 +1,11 @@
+package ain.tolva.nut.backend.alerts;
+
+public class Alerter {
+	
+	AlertType alert;
+	
+	public Alerter(AlertType a) {
+		alert = a;
+	}
+	
+}
